requirements.txt
streamlit
langchain_core
langchain_community
langchain_ollama
pdfplumber
pymupdf
pytesseract
pillow
easyocr
openai-whisper